factor out openssl initialization
