+++ openssl/openssl_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestMain(m *testing.M) {
-	err := openssl.Init()
+	err := openssl.Init("")
 	if err != nil {
 		// An error here could mean that this Linux distro does not have a supported OpenSSL version
 		// or that there is a bug in the Init code.
